requirements.txt
aiohttp
python-dotenv
pytest
requests